@wip: extracting common / util libraries
