+++ service-kit-core/Cargo.toml
@@ -39,6 +39,7 @@ tokio-util = { workspace = true }
 tower-http = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
+service-kit-support = { workspace = true }
 tui-textarea = { workspace = true }
 
 [dev-dependencies]
