+++ service-kit-support/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod assets;
+pub mod errors;
+pub mod storage;
+
+pub type Result<T> = std::result::Result<T, errors::Error>;
